requirements.txt
beautifulsoup4==4.14.3
requests==2.32.5
feedparser==6.0.12
selenium==4.40.0
python-dotenv==1.2.1
